feat: support reordering table
