+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/simple_table/simple_table_more_action.dart
@@ -328,6 +328,11 @@ class _SimpleTableMoreActionMenuState extends State<SimpleTableMoreActionMenu> {
 
   // On mobile, the menu is a action sheet and always shows.
   Widget _buildMobileMenu() {
+    final tableNode = widget.tableCellNode.parentTableNode;
+    if (tableNode == null) {
+      return const SizedBox.shrink();
+    }
+
     return ValueListenableBuilder(
       valueListenable: isShowingMenu,
       builder: (context, isShowingMenu, child) {
@@ -361,7 +366,7 @@ class _SimpleTableMoreActionMenuState extends State<SimpleTableMoreActionMenu> {
           child: SimpleTableMobileDraggableReorderButton(
             index: widget.index,
             type: widget.type,
-            node: widget.tableCellNode,
+            node: tableNode,
             isShowingMenu: this.isShowingMenu,
             editorState: editorState,
             simpleTableContext: simpleTableContext,
